+++ src/linux/q_shlinux.c
@@ -176,7 +176,7 @@ static qboolean CompareAttributes(const char *path, const char *name,
 
 	//return true;
 	  path_len = snprintf(fn, sizeof(fn), "%s/%s", path, name);
-	    if (path_len < 0 || path_len >= sizeof(fn))
+	    if (path_len < 0 || path_len >= (int)sizeof(fn))
 	    {
 	        // Path is too long to stat, treat it as a non-match.
 	        Com_DPrintf("CompareAttributes: Path is too long to stat, skipping: %s/%s\n", path, name);
@@ -228,7 +228,7 @@ char *Sys_FindFirst (const char *path, unsigned musthave, unsigned canhave)
 			    path_len = snprintf (findpath, sizeof(findpath), "%s/%s", findbase, d->d_name);
 
 			    // Check for snprintf errors or truncation.
-			    if (path_len < 0 || path_len >= sizeof(findpath))
+			    if (path_len < 0 || path_len >= (int)sizeof(findpath))
 			    {
 			        // The path was too long and got truncated, or an error occurred.
 			        // Log a warning (optional but good practice) and skip this file.
@@ -260,7 +260,7 @@ char *Sys_FindNext (unsigned musthave, unsigned canhave)
 			    path_len = snprintf (findpath, sizeof(findpath), "%s/%s", findbase, d->d_name);
 
 			    // Check for snprintf errors or truncation.
-			    if (path_len < 0 || path_len >= sizeof(findpath))
+			    if (path_len < 0 || path_len >= (int)sizeof(findpath))
 			    {
 			        Com_DPrintf("Sys_FindNext: Path is too long to process, skipping: %s/%s\n", findbase, d->d_name);
 			        continue;
