+++ src/ref_gl/gl_light.c
@@ -588,7 +588,7 @@ void R_BuildLightMap (const msurface_t *surf, byte *dest, int stride)
 	smax = (surf->extents[0]>>4)+1;
 	tmax = (surf->extents[1]>>4)+1;
 	size = smax*tmax;
-	if (size > (sizeof(s_blocklights)>>4) )
+	if (size > (int)(sizeof(s_blocklights)>>4) )
 		Com_Error (ERR_DROP, "Bad s_blocklights size");
 
 	// set to full bright if no light data
